+++ BubblesDivePlanner/Cylinders/CylinderSetup/GasMixture/IGasMixtureController.cs
@@ -2,6 +2,6 @@ namespace BubblesDivePlanner.Cylinders.CylinderSetup.GasMixture
 {
     public interface IGasMixtureController
     {
-        int CalculateNitrogenMixture(int oxygen, int helium);
+        double CalculateNitrogenMixture(double oxygen, double helium);
     }
 }
