+++ BubblesDivePlanner/Cylinders/CylinderSetup/GasMixture/IGasMixtureModel.cs
@@ -2,8 +2,8 @@ namespace BubblesDivePlanner.Cylinders.CylinderSetup.GasMixture
 {
     public interface IGasMixtureModel
     {
-        int Oxygen { get; set; }
-        int Helium { get; set; }
-        int Nitrogen { get; }
+        double Oxygen { get; set; }
+        double Helium { get; set; }
+        double Nitrogen { get; }
     }
 }
